Store repeated fields as Vectors.
Fixes #58.

For every repeated field `foo`, add a new lens `vector'foo`
whose field type is a `Vector` instead of a list.  Internally,
repeated fields are always stored as `Vectors`, whether or not
the field is actually packed.  Specifically:
- `Data.Vector.Unboxed.Vector` if the field's Haskell representation is
  fixed-length (`Int32`, `Float`, etc.), or
- `Data.Vector.Vector` (i.e., boxed) otherwise (`Text`, `ByteString`, messages,
  etc).

For backwards compatibility, the name `foo` continues to be used
for a list and is implemented as

    foo = vector'foo . iso Vector.toList Vector.fromList

The generated encoding/decoding code uses those `vector'*` lensese
internally.  For parsing, we keep a separate mutable vector for each
repeated field and grow it manually using the
`Data.ProtoLens.Encoding.Growing` module.

Two potentially dodgy things about this change:
1. I couldn't come up with an efficient, simple, and fully safe API
   for interleaving the vector mutation and parsing effects.  I settled
   for exposing `unsafeLiftIO :: IO a -> Parser a` in a separate module
   `Data.ProtoLens.Encoding.Parser.Unsafe`.
   The standard ST functions (e.g. `create` and `createT` from `vector`)
   don't really work here because we have an arbitary set of vectors
   of different element types.
2. In order to get the inner loops to optimize well, I introduced
   a function (born of much testing and staring at Core output):
   ```Data.ProtoLens.Encoding.Bytes.foldMapBuilder :: Vector v a => (a -> Builder) -> v a -> Builder```
   It uses internal Builder functions but produces large speedups
   due to getting GHC to inline the inner loop correctly.  (In fact,
   without it this change is slightly *worse* in encoding performance
   than the version with lists.)

TODO:
- Store enums in unboxed vectors.
- Also store unknown fields in a Vector (boxed) rather than a list.

Hand-tuned foldMapBuilder.